Repos verschoben die ich in falschenordner gemacht habe

+++ src/main/java/de/watchmywatch/repository/storage/AccountRepository.java
@@ -4,5 +4,5 @@ import de.watchmywatch.model.AccountManagment.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Long> {
+public interface AccountRepository extends CrudRepository<Account, Integer> {
 }
